Updated transformations to they are streamlined
This means that all lines are fetched, transformed and immediately dispatched to the destination file.

By doing such thing we achieved memory efficiency increase of 17% on average, and no measurable speed loss. Unfortunately as of now there's no possibility to "add comments" since comments appear to be the same as translations and cannot be distinguished as such.